moved poll, removed ctx params
