+++ pkg/node/connection_pool.go
@@ -1,10 +1,9 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information
 
-package pool
+package node
 
 import (
-	"context"
 	"sync"
 )
 
@@ -23,28 +22,25 @@ func NewConnectionPool() *ConnectionPool {
 }
 
 // Add takes a node ID as the key and a node client as the value to store
-func (mp *ConnectionPool) Add(ctx context.Context, key string, value interface{}) error {
+func (mp *ConnectionPool) Add(key string, value interface{}) error {
 	mp.mu.Lock()
 	defer mp.mu.Unlock()
 	mp.cache[key] = value
-
 	return nil
 }
 
 // Get retrieves a node connection with the provided nodeID
 // nil is returned if the NodeID is not in the connection pool
-func (mp *ConnectionPool) Get(ctx context.Context, key string) (interface{}, error) {
+func (mp *ConnectionPool) Get(key string) (interface{}, error) {
 	mp.mu.Lock()
 	defer mp.mu.Unlock()
-
 	return mp.cache[key], nil
 }
 
 // Remove deletes a connection associated with the provided NodeID
-func (mp *ConnectionPool) Remove(ctx context.Context, key string) error {
+func (mp *ConnectionPool) Remove(key string) error {
 	mp.mu.Lock()
 	defer mp.mu.Unlock()
 	mp.cache[key] = nil
-
 	return nil
 }
