+++ pkg/node/connection_pool_test.go
@@ -1,10 +1,9 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package pool
+package node
 
 import (
-	"context"
 	"sync"
 	"testing"
 
@@ -16,7 +15,6 @@ type TestFoo struct {
 }
 
 func TestGet(t *testing.T) {
-	ctx := context.Background()
 	cases := []struct {
 		pool          *ConnectionPool
 		key           string
@@ -26,7 +24,7 @@ func TestGet(t *testing.T) {
 		{
 			pool: func() *ConnectionPool {
 				p := NewConnectionPool()
-				assert.NoError(t, p.Add(ctx, "foo", TestFoo{called: "hoot"}))
+				assert.NoError(t, p.Add("foo", TestFoo{called: "hoot"}))
 				return p
 			}(),
 			key:           "foo",
@@ -37,7 +35,7 @@ func TestGet(t *testing.T) {
 
 	for i := range cases {
 		v := &cases[i]
-		test, err := v.pool.Get(ctx, v.key)
+		test, err := v.pool.Get(v.key)
 		assert.Equal(t, v.expectedError, err)
 		assert.Equal(t, v.expected, test)
 	}
@@ -64,10 +62,10 @@ func TestAdd(t *testing.T) {
 
 	for i := range cases {
 		v := &cases[i]
-		err := v.pool.Add(context.Background(), v.key, v.value)
+		err := v.pool.Add(v.key, v.value)
 		assert.Equal(t, v.expectedError, err)
 
-		test, err := v.pool.Get(context.Background(), v.key)
+		test, err := v.pool.Get(v.key)
 		assert.Equal(t, v.expectedError, err)
 
 		assert.Equal(t, v.expected, test)
@@ -93,10 +91,10 @@ func TestRemove(t *testing.T) {
 
 	for i := range cases {
 		v := &cases[i]
-		err := v.pool.Remove(context.Background(), v.key)
+		err := v.pool.Remove(v.key)
 		assert.Equal(t, v.expectedError, err)
 
-		test, err := v.pool.Get(context.Background(), v.key)
+		test, err := v.pool.Get(v.key)
 		assert.Equal(t, v.expectedError, err)
 
 		assert.Equal(t, v.expected, test)
